Move root and /samenvatting utils to folder
