+++ apps/melding-form/src/app/(general)/samenvatting/_utils/getContactSummary.ts
@@ -1,0 +1,9 @@
+export const getContactSummary = (label: string, email?: string | null, phone?: string | null) => {
+  if (!email && !phone) return undefined
+
+  return {
+    description: [email, phone].filter((item) => item !== undefined && item !== null), // Filter out undefined or null items
+    key: 'contact',
+    term: label,
+  }
+}
