+++ apps/melding-form/src/app/(general)/samenvatting/_utils/index.ts
@@ -1,0 +1,6 @@
+export { getAdditionalQuestionsSummary } from './getAdditionalQuestionsSummary'
+export { getAttachmentsSummary } from './getAttachmentsSummary'
+export { getContactSummary } from './getContactSummary'
+export { getLocationSummary } from './getLocationSummary'
+export { getMeldingData } from './getMeldingData'
+export { getPrimaryFormSummary } from './getPrimaryFormSummary'
