+++ apps/melding-form/src/app/back-office-entry/route.ts
@@ -5,7 +5,7 @@ import { NextResponse } from 'next/server'
 
 import { client } from '@meldingen/api-client'
 
-import { resolveClassificationRedirect } from '../utils'
+import { resolveClassificationRedirect } from '../_utils/resolveClassificationRedirect'
 import { COOKIES } from '~/constants'
 
 // Configure the API client here, because a route handler does not pass layout.tsx.
